apt_requirement.txt
pybind11
python3-pil
python3-pyqt5
opencv
qt-5
libpcap-dev 
libncurses5-dev 
libprocps-dev 
libxtst-dev 
libxcb-util0-dev 
qttools5-dev-tools 
libdtkwidget-dev 
libdtkwm-dev pkg-config